feat: 删除报告交互优化

+++ src/pages/ReportList.vue
@@ -182,9 +182,31 @@ export default {
     },
 
     async deleteReport(id) {
-      const res = await deleteReport(id);
-      console.log(res);
-      this.getReports(this.pagination.page);
+
+      // 增加 comfirm 交互
+
+      this.$confirm('此操作将永久删除该报告, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(async () => {
+        const res = await deleteReport(id);
+        console.log(res);
+        this.getReports(this.pagination.page);
+        this.$message({
+          type: 'success',
+          message: '删除成功!'
+        });
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });
+      });
+
+      // const res = await deleteReport(id);
+      // console.log(res);
+      // this.getReports(this.pagination.page);
     },
   },
 };
